pyproject.toml
[project]
name = "deepfolder"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
    "python-dotenv>=1.0.0",
    "cryptography>=42.0.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.120.0",
    "itsdangerous>=2.1.0",
    "starlette>=0.40.0",
    "pypdf>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/deepfolder"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.14.0",
]
